+++ Modules/Microsoft365DSC/DSCResources/MSFT_ODSettings/MSFT_ODSettings.psm1
@@ -356,15 +356,15 @@ function Set-TargetResource
     ## https://docs.microsoft.com/en-us/powershell/module/sharepoint-online/set-spotenantsyncclientrestriction?view=sharepoint-ps
     Write-Verbose -Message "Setting other configuration parameters"
     Write-Verbose -Message ($Options | Out-String)
-    if ($Options.ContainsKey("BlockMacSync") -and $Options.ContainsKey("DomainGuids"))
+    if ($Options.ContainsKey("BlockMacSync"))
     {
         Write-Verbose -Message "Updating BlockMacSync {$($Options.BlockMacSync)}"
-        Set-PnPTenantSyncClientRestriction -BlockMacSync:$Options.BlockMacSync -DomainGuids $Options.DomainGuids -Enable:$true | Out-Null
+        Set-PnPTenantSyncClientRestriction -BlockMacSync:$Options.BlockMacSync 
     }
-    elseif ($Options.ContainsKey("DomainGuids") -and ($Options.ContainsKey("BlockMacSync") -eq $false))
+    if ($Options.ContainsKey("DomainGuids"))
     {
         Write-Verbose -Message "Updating DomainGuids"
-        Set-PnPTenantSyncClientRestriction -DomainGuids $Options.DomainGuids -Enable:$true| Out-Null
+        Set-PnPTenantSyncClientRestriction -DomainGuids $Options.DomainGuids -Enable:$true 
     }
 
     if ($Options.ContainsKey("ExcludedFileExtensions"))
@@ -376,19 +376,20 @@ function Set-TargetResource
             $BlockedFileTypes += $fileTypes + ';'
         }
 
-        Set-PnPTenantSyncClientRestriction -ExcludedFileExtensions $BlockedFileTypes | Out-Null
+        Set-PnPTenantSyncClientRestriction -ExcludedFileExtensions $BlockedFileTypes  
     }
     if ($Options.ContainsKey("DisableReportProblemDialog"))
     {
         Write-Verbose -Message "Updating DisableReportProblemDialog"
-        Set-PnPTenantSyncClientRestriction -DisableReportProblemDialog:$Options.DisableReportProblemDialog | Out-Null
+        Set-PnPTenantSyncClientRestriction -DisableReportProblemDialog:$Options.DisableReportProblemDialog  
     }
 
     if ($Options.ContainsKey("GrooveBlockOption"))
     {
         Write-Verbose -Message "Updating GrooveBlockOption"
-        Set-PnPTenantSyncClientRestriction -GrooveBlockOption $Options.GrooveBlockOption | Out-Null
+        Set-PnPTenantSyncClientRestriction -GrooveBlockOption $Options.GrooveBlockOption  
     }
+ 
 }
 
 function Test-TargetResource